* Renamed iq_discrimination.py to discriminators.py
