+++ qiskit/ignis/measurement/discriminator/discriminators.py
